handle corrupt hvm headers state when "prev" block header is missing (#31)
* handle corrupt hvm headers state when "prev" block header is missing

when we insert a block header into our "headers only" tbc instance, it performs a check to ensure that the previous block header is in the database.  otherwise it errors.  this error is a "not found" error.  if we come accross this, we know the headers data store is corrupt and must be fully restored. so do so.

* change log to warning

* Update docker.yml